TradeBotV72/TradeBotV72/requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
ccxt>=4.0.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
tabulate>=0.9.0